+++ lib/components/z-escala.dart
@@ -7,11 +7,11 @@ import 'package:z_components/view-model/EscalaViewModel.dart';
 
 class ZEscala extends StatefulWidget {
   final String token;
-  Key key;
+  final Key key;
+  final ValueChanged<ZCollectionItem> onChange;
 
 
-  ZEscala({  Key key
-    ,this.token}) : super(key: key);
+  ZEscala({this.key, @required this.token, this.onChange});
 
   @override
   State<StatefulWidget> createState() => ZEscalaState();
@@ -21,8 +21,6 @@ class ZEscalaState extends State<ZEscala> {
   ZCollectionItem _itemSelecionado;
 
   ZCollectionItem get itemSelecionado => _itemSelecionado;
-
-  var key = new GlobalKey<ZCollectionState>();
 
   IZEscalaService _service;
 
@@ -38,15 +36,15 @@ class ZEscalaState extends State<ZEscala> {
   @override
   Widget build(BuildContext context) {
     return new ZCollection(
-      key: key,
       titulo: "Escalas",
       lista: _escalas
           .map((x) => new ZCollectionItem(
-          chave: x.idEscala, titulo: x.nome, valor: x.nome))
+              chave: x.idEscala, titulo: x.nome, valor: x.nome))
           .toList(),
       onChange: (item) {
         _itemSelecionado = item;
-      },
+        if(widget.onChange != null) widget.onChange(item);
+      }
     );
   }